objectified-rest/pyproject.toml
[project]
name = "objectified-rest"
version = "1.0.8"
description = "REST API for the Objectified platform; CRUD scaffolding from database design"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Objectified Team"}
]
keywords = ["fastapi", "openapi", "rest-api", "pydantic", "postgresql"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.31.0",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.5.2",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
    "pyjwt>=2.8.0",
    "pyyaml>=6.0.2",
    "argon2-cffi>=23.1.0",
    "bcrypt>=4.0.0",
    "jsonschema>=4.25.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.6",
]
otel = [
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-exporter-otlp-proto-http>=1.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
